Create scoreSheet component
