+++ src/assets/home.scss
@@ -60,6 +60,18 @@
         width: 45px;
         padding: 0;
       }
+
+      /* Remove Scroll in Chrome, Safari, Edge, Opera */
+      input::-webkit-outer-spin-button,
+      input::-webkit-inner-spin-button {
+        -webkit-appearance: none;
+        margin: 0;
+      }
+
+      /* Remove Scroll in Firefox */
+      input[type=number] {
+        -moz-appearance: textfield;
+      }
     }
 
     .team2 {
