+++ src/helpers/displayHelpers.js
@@ -1,3 +1,5 @@
+import GameInput from '../components/gameInput';
+
 const displayHeaders = (headers) => (
   <tr>
     {headers.map((header) => (
@@ -44,79 +46,22 @@ const displayTeams = (teamList) => {
   return <tr />;
 };
 
-const addScoreToGame = (currGame, value, side) => {
-  const games = JSON.parse(localStorage.games);
-
-  const thisGame = games.find((elem) => elem.id === currGame);
-  thisGame[side] = value;
-
-  localStorage.games = JSON.stringify(games);
-
-  if (Object.values(thisGame)[1] !== '' && Object.values(thisGame)[2] !== '') {
-    const winner = Object.keys(games[currGame]).splice(1, 2)
-      .reduce((a, b) => (games[currGame][a] > games[currGame][b] ? a : b));
-    const loser = Object.keys(games[currGame]).splice(1, 2)
-      .reduce((a, b) => (games[currGame][a] < games[currGame][b] ? a : b));
-
-    const teamList = JSON.parse(localStorage.teamList);
-    if (winner === loser) {
-      let otherTeam = Object.keys(games[currGame]);
-
-      otherTeam = otherTeam.filter((n) => !['id', winner].includes(n));
-      const firstTeam = teamList.find((team) => team.Team === winner);
-      const secondTeam = teamList.find((team) => team.Team === otherTeam[0]);
-      firstTeam.Played += 1;
-      firstTeam.Points += 1;
-      firstTeam.Draw += 1;
-      secondTeam.Played += 1;
-      secondTeam.Points += 1;
-      secondTeam.Draw += 1;
-    } else {
-      const winningTeam = teamList.find((team) => team.Team === winner);
-      winningTeam.Played += 1;
-      winningTeam.Win += 1;
-      winningTeam.Points += 3;
-      const losingTeam = teamList.find((team) => team.Team === loser);
-      losingTeam.Played += 1;
-      losingTeam.Loss += 1;
-    }
-    localStorage.teamList = JSON.stringify(teamList);
-  }
-};
-
-const displayScores = (games, dispatch, getGames, getTeams) => {
-  const handleKeyPress = (e, game, side) => {
-    if (e.key === 'Enter' || e.keycode === 13) {
-      addScoreToGame(game, e.target.value, side);
-      dispatch(getGames());
-      dispatch(getTeams());
-    }
-  };
-  return (
-    games.map((game) => (
+const displayScores = (games, handleKeyPress) => (
+  games.map((game) => {
+    const [team1, team2] = Object.keys(game).splice(1, 2);
+    const [team1Score, team2Score] = Object.values(game).splice(1, 2);
+    return (
       <>
-        <span className="team1">{(Object.keys(game)[1])}</span>
-        {Object.values(game)[1] === '' || Object.values(game)[2] === ''
-          ? (
-            <span className="score-input">
-              <input
-                type="text"
-                onKeyPress={(e) => handleKeyPress(e, game.id, Object.keys(game)[1])}
-                placeholder="0"
-              />
-              {' : '}
-              <input
-                type="text"
-                onKeyPress={(e) => handleKeyPress(e, game.id, Object.keys(game)[2])}
-                placeholder="0"
-              />
-            </span>
-          )
-          : <span className="score">{`${(Object.values(game)[1])} : ${(Object.values(game)[2])}`}</span>}
-        <span className="team2">{(Object.keys(game)[2])}</span>
+        <span className="team1">{team1}</span>
+
+        {team1Score === '' || team2Score === ''
+          ? <GameInput onEnter={handleKeyPress} gameID={game.id} team1={team1} team2={team2} />
+          : <span className="score">{`${team1Score} : ${(team2Score)}`}</span>}
+
+        <span className="team2">{team2}</span>
       </>
-    ))
-  );
-};
+    );
+  })
+);
 
 export { displayHeaders, displayTeams, displayScores };
